move tpcc to other workloads
