+++ ydb/library/workload/tpcc/ut/timer_queue_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/public/lib/ydb_cli/commands/tpcc/timer_queue.h>
+#include <ydb/library/workload/tpcc/timer_queue.h>
 
 #include <library/cpp/testing/unittest/registar.h>
 #include <library/cpp/testing/gmock_in_unittest/gmock.h>
