+++ ydb/library/workload/ya.make
@@ -21,6 +21,7 @@ RECURSE(
     log
     mixed
     stock
+    tpcc
     tpc_base
     tpcds
     tpch
