--- ydb/public/lib/ydb_cli/commands/tpcc/command_root.cpp
@@ -1,13 +1,0 @@
-#include "command_root.h"
-
-#include "command_run.h"
-
-namespace NYdb::NConsoleClient {
-
-TCommandTPCC::TCommandTPCC()
-    : TClientCommandTree("tpcc", {}, "YDB TPC-C workload")
-{
-    AddCommand(std::make_unique<TCommandTPCCRun>());
-}
-
-} // namespace NYdb::NConsoleClient
