+++ ydb/public/lib/ydb_cli/commands/ya.make
@@ -35,6 +35,7 @@ SRCS(
     ydb_tools.cpp
     ydb_workload.cpp
     ydb_workload_import.cpp
+    ydb_workload_tpcc.cpp
     ydb_yql.cpp
 )
 
@@ -53,7 +54,6 @@ PEERDIR(
     ydb/public/lib/ydb_cli/commands/interactive
     ydb/public/lib/ydb_cli/commands/sdk_core_access
     ydb/public/lib/ydb_cli/commands/topic_workload
-    ydb/public/lib/ydb_cli/commands/tpcc
     ydb/public/lib/ydb_cli/commands/transfer_workload
     ydb/public/lib/ydb_cli/commands/ydb_discovery
     ydb/public/lib/ydb_cli/common
