+++ ydb/public/lib/ydb_cli/commands/ydb_workload_tpcc.cpp
@@ -1,17 +1,39 @@
-#include "command_run.h"
+#include "ydb_workload_tpcc.h"
 
-#include "runner.h"
+#include <ydb/library/workload/tpcc/runner.h>
+
+#include <ydb/public/lib/ydb_cli/commands/ydb_command.h>
 
 #include <util/system/info.h>
 
 namespace NYdb::NConsoleClient {
 
 namespace {
-    constexpr int DEFAULT_WAREHOUSE_COUNT = 1;
-    constexpr int DEFAULT_WARMUP_SECONDS = 60; // TODO
-    constexpr int DEFAULT_RUN_SECONDS = 60; // TODO
 
-} // anonymous
+//-----------------------------------------------------------------------------
+
+constexpr int DEFAULT_WAREHOUSE_COUNT = 1;
+constexpr int DEFAULT_WARMUP_SECONDS = 60; // TODO
+constexpr int DEFAULT_RUN_SECONDS = 60; // TODO
+
+//-----------------------------------------------------------------------------
+
+class TCommandTPCCRun
+    : public TYdbCommand
+{
+public:
+    TCommandTPCCRun();
+    ~TCommandTPCCRun();
+
+    virtual void Config(TConfig& config) override;
+    virtual void Parse(TConfig& config) override;
+    virtual int Run(TConfig& config) override;
+
+private:
+    int WarehouseCount;
+    int WarmupSeconds;
+    int RunSeconds;
+};
 
 TCommandTPCCRun::TCommandTPCCRun()
     : TYdbCommand("run", {}, "run benchmark")
@@ -51,6 +73,16 @@ int TCommandTPCCRun::Run(TConfig& config) {
     NTPCC::RunSync(tpccConfig);
 
     return 0;
+}
+
+} // anonymous
+
+//-----------------------------------------------------------------------------
+
+TCommandTPCC::TCommandTPCC()
+    : TClientCommandTree("tpcc", {}, "YDB TPC-C workload")
+{
+    AddCommand(std::make_unique<TCommandTPCCRun>());
 }
 
 } // namespace NYdb::NConsoleClient
