+++ ydb/public/lib/ydb_cli/commands/ydb_workload_tpcc.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <ydb/public/lib/ydb_cli/commands/ydb_command.h>
-#include <ydb/public/lib/ydb_cli/common/interruptible.h>
 
 namespace NYdb::NConsoleClient {
